move lemon toast component into subfolder
