+++ frontend/src/lib/components/ExportButton/exporter.tsx
@@ -2,7 +2,7 @@ import { AnimationType } from 'lib/animations/animations'
 import api from 'lib/api'
 import { Animation } from 'lib/components/Animation/Animation'
 import { dayjs } from 'lib/dayjs'
-import { lemonToast } from 'lib/lemon-ui/lemonToast'
+import { lemonToast } from 'lib/lemon-ui/LemonToast/LemonToast'
 import { Spinner } from 'lib/lemon-ui/Spinner/Spinner'
 import { delay } from 'lib/utils'
 import posthog from 'posthog-js'
