+++ frontend/src/lib/components/ObjectTags/objectTagsLogic.ts
@@ -1,6 +1,6 @@
 import equal from 'fast-deep-equal'
 import { actions, kea, key, listeners, path, props, propsChanged, reducers, selectors } from 'kea'
-import { lemonToast } from 'lib/lemon-ui/lemonToast'
+import { lemonToast } from 'lib/lemon-ui/LemonToast/LemonToast'
 
 import type { objectTagsLogicType } from './objectTagsLogicType'
