+++ frontend/src/lib/lemon-ui/LemonToast/LemonToast.tsx
@@ -3,7 +3,7 @@ import { Spinner } from 'lib/lemon-ui/Spinner/Spinner'
 import posthog from 'posthog-js'
 import { toast, ToastContentProps as ToastifyRenderProps, ToastOptions } from 'react-toastify'
 
-import { LemonButton } from './LemonButton'
+import { LemonButton } from '../LemonButton'
 
 export function ToastCloseButton({ closeToast }: { closeToast?: () => void }): JSX.Element {
     return (
