+++ frontend/src/lib/lemon-ui/LemonToast/index.ts
@@ -1,0 +1,1 @@
+export { lemonToast } from './LemonToast'
