+++ frontend/src/scenes/dashboard/newDashboardLogic.ts
@@ -3,7 +3,7 @@ import { forms } from 'kea-forms'
 import { router } from 'kea-router'
 import api from 'lib/api'
 import { DashboardRestrictionLevel } from 'lib/constants'
-import { lemonToast } from 'lib/lemon-ui/lemonToast'
+import { lemonToast } from 'lib/lemon-ui/LemonToast/LemonToast'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { teamLogic } from 'scenes/teamLogic'
 import { urls } from 'scenes/urls'
