+++ frontend/src/scenes/data-management/definition/definitionEditLogic.ts
@@ -2,7 +2,7 @@ import { beforeUnmount, connect, kea, key, path, props } from 'kea'
 import { forms } from 'kea-forms'
 import { loaders } from 'kea-loaders'
 import api from 'lib/api'
-import { lemonToast } from 'lib/lemon-ui/lemonToast'
+import { lemonToast } from 'lib/lemon-ui/LemonToast/LemonToast'
 import { capitalizeFirstLetter } from 'lib/utils'
 import {
     definitionLogic,
