+++ frontend/src/scenes/organization/ConfirmOrganization/confirmOrganizationLogic.ts
@@ -2,7 +2,7 @@ import { actions, kea, path, reducers } from 'kea'
 import { forms } from 'kea-forms'
 import { urlToAction } from 'kea-router'
 import api from 'lib/api'
-import { lemonToast } from 'lib/lemon-ui/lemonToast'
+import { lemonToast } from 'lib/lemon-ui/LemonToast/LemonToast'
 
 import type { confirmOrganizationLogicType } from './confirmOrganizationLogicType'
