+++ frontend/src/scenes/plugins/source/pluginSourceLogic.tsx
@@ -4,7 +4,7 @@ import { loaders } from 'kea-loaders'
 import { beforeUnload } from 'kea-router'
 import api from 'lib/api'
 import { FormErrors } from 'lib/forms/Errors'
-import { lemonToast } from 'lib/lemon-ui/lemonToast'
+import { lemonToast } from 'lib/lemon-ui/LemonToast/LemonToast'
 import { validateJson } from 'lib/utils'
 import { frontendAppsLogic } from 'scenes/apps/frontendAppsLogic'
 import { pluginsLogic } from 'scenes/plugins/pluginsLogic'
