+++ frontend/src/scenes/settings/user/personalAPIKeysLogic.ts
@@ -1,7 +1,7 @@
 import { kea, listeners, path } from 'kea'
 import { loaders } from 'kea-loaders'
 import api from 'lib/api'
-import { lemonToast } from 'lib/lemon-ui/lemonToast'
+import { lemonToast } from 'lib/lemon-ui/LemonToast/LemonToast'
 import { copyToClipboard } from 'lib/utils/copyToClipboard'
 
 import { PersonalAPIKeyType } from '~/types'
